+++ CHANGELOG.md
@@ -9,11 +9,16 @@
 - Add single layer function `gfo.warp` to warp geometries based on GCP's (#151)
 - Add (private) function to calculate view angles from a point towards a GeoDataFrame (#140)
 - Add (private) function to calculate topologic simplify (#146)
-- Small changes to support geopandas 0.11+, newer pandas versions,... (#143, #147, #153)
+- Add (private) is_valid_reason function for geodataframe (#164)
+- Small changes to support geopandas 0.11+, newer pandas versions,... (#143, #147, #153, #163)
 
 ### Bugs fixed
 
 - Fix typo in hardcoded 31370 custom prj string (#142)
+
+### Deprecations and compatibility notes
+
+- Make topojson an optional dependency (#165)
 
 ## 0.5.0 (2022-06-08)