+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Bugs fixed
 
+- Fix that groupby columns in dissolve were not treated case insensitive (#162)
 
 ### Deprecations and compatibility notes